Fixed warnings from project. Fixed bug with declaring integer as pointer at Sentegrity_SubClassresult_Object.

+++ Sentegrity_Combined/Core Detection/TrustScore Computation/Sentegrity_TrustScore_Computation.m
@@ -532,7 +532,7 @@
                 Sentegrity_SubClassResult_Object *subClassResultObject = [[Sentegrity_SubClassResult_Object alloc] init];
                 
                 // Determine level of trust in the subclass by dividing current amount by total possible
-                int percentOfTrust;
+                NSInteger percentOfTrust;
                 
                 // Should always be > 0 otherwise must be a policy error
                 if(subClass.totalPossibleScore > 0){
